refactor: move ThreadUtils to internal
