+++ clients/src/main/java/org/apache/kafka/common/utils/internals/ThreadUtils.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.kafka.common.utils;
+package org.apache.kafka.common.utils.internals;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
