+++ coordinator-common/src/main/java/org/apache/kafka/coordinator/common/runtime/CoordinatorBackgroundThreadPoolExecutor.java
@@ -16,7 +16,7 @@
  */
 package org.apache.kafka.coordinator.common.runtime;
 
-import org.apache.kafka.common.utils.ThreadUtils;
+import org.apache.kafka.common.utils.internals.ThreadUtils;
 import org.apache.kafka.common.utils.Time;
 
 import java.util.Objects;
