+++ server-common/src/main/java/org/apache/kafka/server/util/timer/SystemTimer.java
@@ -17,7 +17,7 @@
 package org.apache.kafka.server.util.timer;
 
 import org.apache.kafka.common.utils.internals.KafkaThread;
-import org.apache.kafka.common.utils.ThreadUtils;
+import org.apache.kafka.common.utils.internals.ThreadUtils;
 import org.apache.kafka.common.utils.Time;
 
 import java.util.concurrent.DelayQueue;
