+++ trogdor/src/main/java/org/apache/kafka/trogdor/workload/ExternalCommandWorker.java
@@ -18,7 +18,7 @@
 package org.apache.kafka.trogdor.workload;
 
 import org.apache.kafka.common.internals.KafkaFutureImpl;
-import org.apache.kafka.common.utils.ThreadUtils;
+import org.apache.kafka.common.utils.internals.ThreadUtils;
 import org.apache.kafka.trogdor.common.JsonUtil;
 import org.apache.kafka.trogdor.common.Platform;
 import org.apache.kafka.trogdor.task.TaskWorker;
