+++ trogdor/src/test/java/org/apache/kafka/trogdor/task/SampleTaskWorker.java
@@ -18,7 +18,7 @@
 package org.apache.kafka.trogdor.task;
 
 import org.apache.kafka.common.internals.KafkaFutureImpl;
-import org.apache.kafka.common.utils.ThreadUtils;
+import org.apache.kafka.common.utils.internals.ThreadUtils;
 import org.apache.kafka.trogdor.common.Platform;
 
 import com.fasterxml.jackson.databind.node.TextNode;
